refactor() move files to core and remove legacy variables
