+++ src/database/dml.py
@@ -33,9 +33,9 @@ from ..core.constants import TABLES_INFO_DICT
 from ..core.schemas import TableInfo
 from ..utils.model_utils import get_table_columns
 from .schemas import Database
-from ..utils.file_loader.file_loader import FileLoader
-from ..utils.file_loader.uploader import async_upsert
-from ..utils.file_loader.connection_factory import (
+from ..core.loading.file_loader.file_loader import FileLoader
+from ..core.loading.file_loader.uploader import async_upsert
+from ..core.loading.file_loader.connection_factory import (
     create_asyncpg_pool_from_sqlalchemy,
     extract_primary_keys,
     get_column_types_mapping
